ogit-clone: a little more cleanup
- Split http cloning details out into clone_http
- Static'ize functions that don't need global linkage
