+++ Makefile
@@ -6,7 +6,8 @@ PROG=	ogit
 
 SRCS=	ogit.c lib/ini.c lib/index.c lib/common.c lib/pack.c remote.c init.c \
 	lib/zlib-handler.c lib/buffering.c \
-	hash-object.c update-index.c cat-file.c log.c clone.c index-pack.c
+	hash-object.c update-index.c cat-file.c log.c clone.c clone_http.c \
+	index-pack.c
 
 CLEANFILES+=	${PROG}.core
